Moved search models to separate file
